+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+- Fixed Discover tab lag and stutter
+
 - Added Background Refresh and Automated Reminder Add
 
 ## [2.2.0] - 2026-09-09